BackEnd/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
asgiref=3.5.2=pyhd8ed1ab_0
backports=1.0=py_2
backports.zoneinfo=0.2.1=py38h0dd4459_5
brotlipy=0.7.0=py38hed1de0f_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.6.15=h033912b_0
certifi=2022.6.15=py38h50d1736_0
cffi=1.15.1=py38h86886aa_0
chardet=3.0.4=py38h5347e94_1008
charset-normalizer=2.1.0=pyhd8ed1ab_0
click=8.1.3=py38h50d1736_0
cryptography=37.0.4=py38h0f8513e_0
django=4.0.6=py38h50d1736_0
django-cors-headers=3.13.0=pyhd8ed1ab_0
django-filter=22.1=pyhd8ed1ab_0
djangorestframework=3.13.1=pyhd8ed1ab_0
idna=2.8=py38_1000
krb5=1.19.3=hb49756b_0
libcxx=14.0.6=hce7ea42_0
libedit=3.1.20191231=h0678c8f_2
libffi=3.4.2=h0d85af4_5
libpq=14.4=hf6bb32a_0
libzlib=1.2.12=hfe4f2af_2
ncurses=6.3=h96cf925_1
openssl=1.1.1q=hfe4f2af_0
pip=22.1.2=pyhd8ed1ab_0
psycopg2=2.9.3=py38h05cbfbd_0
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py38h50d1736_5
python=3.8.13=h394c593_0_cpython
python-dotenv=0.20.0=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytz=2022.1=pyhd8ed1ab_0
readline=8.1.2=h3899abd_0
requests=2.28.1=pyhd8ed1ab_0
setuptools=63.2.0=py38h50d1736_0
sqlite=3.39.1=hd9f0692_0
sqlparse=0.4.2=pyhd8ed1ab_0
tk=8.6.12=h5dbffcc_0
typing_extensions=4.3.0=pyha770c72_0
tzdata=2022a=h191b570_0
urllib3=1.25.11=py_0
wheel=0.37.1=pyhd8ed1ab_0
xz=5.2.5=haf1e3a3_1
zlib=1.2.12=hfe4f2af_2
